tectonic/xetex-synctex.[ch]: rename from synctex.[ch]
