+++ build.rs
@@ -150,7 +150,6 @@ fn main() {
         .file("tectonic/dpx-type1c.c")
         .file("tectonic/dpx-unicode.c")
         .file("tectonic/dpx-vf.c")
-        .file("tectonic/synctex.c")
         .file("tectonic/texmfmp.c")
         .file("tectonic/xetex0.c")
         .file("tectonic/XeTeX_ext.c")
@@ -165,6 +164,7 @@ fn main() {
         .file("tectonic/xetex-scaledmath.c")
         .file("tectonic/xetex-shipout.c")
         .file("tectonic/xetex-stringpool.c")
+        .file("tectonic/xetex-synctex.c")
         .define("HAVE_ZLIB", "1")
         .define("HAVE_ZLIB_COMPRESS2", "1")
         .define("ZLIB_CONST", "1")
