+++ tectonic/xetex-synctex.c
@@ -6,7 +6,7 @@
 #include "xetex-core.h"
 #include "tectonic.h"
 #include "xetexd.h"
-#include "synctex.h"
+#include "xetex-synctex.h"
 #include "core-bridge.h"
 
 #include <stdio.h>
